3/10/24

+++ src/client/components/plan/PlanDrawer.jsx
@@ -4,24 +4,32 @@ import EditIcon from '@mui/icons-material/Edit';
 import DeleteIcon from '@mui/icons-material/Delete';
 import axios from 'axios';
 import { HostName } from '../../script/HostName';
-import {goodAlert,badAlert} from '../../script/sweet';
-
+import {goodAlert,badAlert,deleteConfirm} from '../../script/sweet';
+import { useSwitch } from '../../context/SwitchContext';
 
 
 const PlanDrawer = ({ open, onClose, plans, selectedDate, onAddPlan, setIsEdit, setSelectedPlan, setPopupOpen }) => {
+  const {toggleSwitch} = useSwitch();
   const setToEdit = (plan) => {
     setIsEdit(true);
     setSelectedPlan(plan);
     setPopupOpen(true);
+    toggleSwitch();
   };
   
 
   const handleDeletePlan = async (plan) => {
+    const result = await deleteConfirm('Are you sure you want to delete this plan?', 'This action cannot be undone.');
     try {
+     if(result){
       const response = await axios.delete(`${HostName}/api/plans/delete/${plan.id}`, { withCredentials: true });
       if(response.status === 200){
         goodAlert('Plan deleted successfully');
+        toggleSwitch();
       }
+     }
+     return;
+   
     } catch (error) {
       badAlert('Error deleting plan:', error.message);
       console.error('Error deleting plan:', error);
